pyproject.toml
[project]
name = "mediatools"
version = "0.2.0"
description = "A simple CLI tool to convert audio and image files"
authors = [
    { name = "Maria Matyushenko", email = "marimatush@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "pydub",
    "pillow"
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.flake8]
max-line-length = 100
